introduces process links to save memory
The previous implementation of goprocess lost lots of memory,
as dead processes were being held indefinitely, particularly
in the main parent. This commit introduces processLinks, an
abstraction that binds two processes together.

It is not multiple-implementation friendly. Right now it is
a hack that works. The introduction of this datastructure
can significantly simplify other things, and thus lots of
refactoring could be done. I want to get this out fast for
IPFS as lots of memory is being lost.